Use 32 bit integer for argument to memset

+++ lib/CodeGen/CGNonTrivialStruct.cpp
@@ -653,7 +653,7 @@ struct GenDefaultInitialize
       return;
     }
 
-    llvm::Constant *SizeVal = CGF->Builder.getInt64(Size.getQuantity());
+    llvm::Constant *SizeVal = CGF->Builder.getInt32(Size.getQuantity());
     Address DstAddr = getAddrWithOffset(Addrs[DstIdx], CurStackOffset, FD);
     Address Loc = CGF->Builder.CreateElementBitCast(DstAddr, CGF->Int8Ty);
     CGF->Builder.CreateMemSet(Loc, CGF->Builder.getInt8(0), SizeVal,
